clientws/requirements.txt
# REST API Generation
fastapi==0.85.0
uvicorn==0.18.2
pydantic==1.9.1
SQLAlchemy==1.4.39
aiosqlite==0.17.0
coloredlogs==15.0.1
PyYAML==6.0
pika==1.2.0
flask==3.0.0
PyJWT==2.8.0
cryptography